feat(NetworkMonitor): refactor snapshot handling and enhance duration display in NetworkMonitorWindow

+++ src/main/java/me/mdbell/awtea/util/ui/Theme.java
@@ -144,7 +144,7 @@ public class Theme {
 			return "?";
 		}
 		if (bytes < 1024) {
-			return bytes + " B";
+			return ((int)bytes) + " B";
 		}
 
 		int unit = 1024;
